==> 7-coerce-to-boolean.ipynb <==
# Coercing to Booleans

### Introduction

In this lesson, we'll work through identifying and coercing data to boolean values.  This will also prepare us to identify and coerce categorical values in our dataset.

### Loading our AirBnb Data

For this lesson, we'll work with [AirBnb listings in Berlin](https://www.kaggle.com/brittabettendorf/berlin-airbnb-data).  Let's load our data.

import pandas as pd
url = "https://raw.githubusercontent.com/jigsawlabs-student/engineering-large-datasets/master/nums_and_dates_ten_k.csv"
df = pd.read_csv(url, index_col = 0)

df.select_dtypes('object').shape

potential_date_cols = ['last_scraped',
 'host_since',
 'calendar_last_scraped',
 'first_review',
 'last_review']
df[potential_date_cols] = df[potential_date_cols].astype('datetime64')

df.shape

Lucky for us, we already have our good amount of our data already coerced.  But we still have more work to do.

object_df = df.select_dtypes(include = 'object')

object_df.shape

def contains_date(column):
#     remove nas first, potentially use all
    regex_string = (r'^\d{1,2}-\d{1,2}-\d{4}$|^\d{4}-\d{1,2}-\d{1,2}$' + 
'|^\d{1,2}\/\d{1,2}\/\d{4}$|^\d{4}\/\d{1,2}\/\d{1,2}$')
    return column.str.contains(regex_string).any()

### Feature engineering

So a lot of our columns are still of type object.  Let's take a look at some of our object columns. 

# object_df.dtypes

Where a larger percentage of the values in our columns repeat, we can think of them as categorical, and eventually one hot encode them.  So we wrote a method called `percent_different` that returns the percent of unique values that make up a series.  If most of the values in a series are unique, then it is not a categorical column.  

So in `find_categorical`, we loop through our columns, identifying those where `percent_different` is not too large - and those are our categorical columns.

def percent_different(df_series):
    series_filled = df_series.dropna()
    return len(series_filled.unique())/len(series_filled)

def find_categorical(df, threshold = .5):    
    categorical_df = pd.DataFrame({})
    for column in df.columns:
        if percent_different(df[column]) < threshold:
            categorical_df[column] = df[column]
    return categorical_df 

Let's see how this works.

potential_cat = find_categorical(object_df)

potential_cat.shape

potential_cat[:2]

It looks like it did a good job.

### Combine with Selecting Categorical Columns

The next step is to take a look at the values in those identified columns, to see if indeed they are full of categories.  Our `get_multiple_val_counts` method loops through a dataframe, providing the top `value_counts` values, and the related column.

def get_multiple_val_counts(df, num_vals = 1):
    return [df[column].value_counts(normalize=True).iloc[:num_vals] for column in df.columns]

get_multiple_val_counts(potential_cat)[:3]

And summarize cats, puts this information in an easier to work with numpy array.

import numpy as np
def summarize_cats(df):
    multiple_val_counts = get_multiple_val_counts(df)
    stacked_counts = np.vstack([np.array([val_count.name, val_count.index[0], float(val_count.values[0])]) for val_count in multiple_val_counts])
    sorted_cols = np.argsort(stacked_counts.reshape(-1, 3)[:, 2].astype('float'))
    return stacked_counts[sorted_cols[::-1]]

summary = summarize_cats(potential_cat)
summary[:10]

The first column in summary is the name of the column, the second is the top value, and the last column is the percent of the column the value was in. 

### Identifying Boolean Values

From the summary grid, we can start to see some strings that are really boolean values.  These are the columns with `t` or `f` as their top values.  

summary[:3]

Let's select all of the columns from our summary that have values of `t` or `f`.

boolean_summary = summary[np.isin(summary[:, 1], ['t', 'f'])]

true_boolean_cols = boolean_summary[:, 0]
true_boolean_cols

From here, we see that some of these columns have `na` values.

potential_cat[true_boolean_cols].isna().sum()/potential_cat.shape[0]

We'll have to coerce these values as well.  

> It's a small enough percentage that we can just coerce them to false values.

Now we'll use our MissingIndicator to convert these columns to have True or False values.  We do so, we by having the transformer set `t` to True, and all other values to False.  We'll also coerce the nan values to `f` beforehand.

> We can loop through to do this for each of our boolean columns.

from sklearn.impute import SimpleImputer, MissingIndicator
steps = [([col], [SimpleImputer(strategy = 'constant', missing_values= np.nan, fill_value = 'f'),
                  MissingIndicator(missing_values = 't')
                 ]
         ) 
         for col in true_boolean_cols]

And then place these steps in a DataFrameMapper.

from sklearn_pandas import DataFrameMapper
boolean_mapper = DataFrameMapper(steps, df_out = True)

# df[bools_df.columns]

bool_df = boolean_mapper.fit_transform(df)

bool_df[:2]

Then we can update our dataframe.

df.loc[:, bool_df.columns] = bool_df

df.select_dtypes('object').shape

And store our dataframe.

df.to_csv('./listings_coerced_booleans.csv')

And finally, we can store our datatypes.

import json
data = df.dtypes.astype(str).to_dict()

file = './coerced_bools_dtypes.json'

with open(file, 'w') as f:
    json.dump(data, f)

Then we can use numpy to identify our remaining potential_cat columns that we should coerce.

import numpy as np
remaining_cat_cols = np.setdiff1d(potential_cat.columns, bool_df.columns)

remaining_cat_cols

### Summary

In this lesson, we were introduced to some of the methods for handling boolean and categorical data.  We saw that we identified our categorical columns by looking at the percent different.  If not a large percent of a column's values are different, it is likely categorical or boolean.  We then used our `summarize_cats` method to view the top values in each of the columns, along with how often they occur.

Finally, we used the `MissingImputer` to convert values in almost boolean columns to True and False values.
