==> 98-coercing-missing-values.ipynb <==
# Coercing Missing Values

### Introduction

In this lesson, let's apply our strategy of replacing missing values with the mean.

import pandas as pd
X_train = pd.read_feather('X_train_listings.feather')

na_cols = X_train.select_dtype(exclude = '')

isna().any(axis = 0).index

# na_cols

imputers = [([col], SimpleImputer()) for col in na_cols]
    

mapper = DataFrameMapper(imputers)

# mapper.fit_transform(X_train)

from sklearn_pandas import DataFrameMapper
from sklearn.impute import SimpleImputer
import numpy as np

replace_zero_cols = ['availability_30', 'availability_60', 'availability_90']

mapper = DataFrameMapper([
    (['availability_30'], SimpleImputer(missing_values = 0, 
                                      strategy = 'constant',
                                      fill_value = np.nan)),
    (['availability_60'], SimpleImputer(missing_values = 0, 
                                      strategy = 'constant',
                                      fill_value = np.nan)),
    (['availability_90'], SimpleImputer(missing_values = 0, 
                                      strategy = 'constant',
                                      fill_value = np.nan)),
    (['maximum_nights'], SimpleImputer(missing_values = 999, 
                                      strategy = 'constant',
                                      fill_value = np.nan))
], df_out = True)
